requirements.txt
wyoming==1.5.3
nemo_toolkit[asr]
resemblyzer==0.1.3
numpy>=1.21.0
scipy>=1.7.0
torch
librosa
soundfile
